Trees_idle_cutting_door
So In this one the unvroken door now has an animation, just like the broken one. All 5 trees have 2 animation 1 idle and one vibrating to be used when its being attacked. All 5 trees have the correct script, is on the correct layer, and have a working animation tree. They are all in the folder with the long name, and ready to be draged out. The unbroken door have been placed in the right spot so no need to drag that one out.