New_timber_drop
So now the timber has a new animation. That can turn 360 degrees. Its better done, and can now be moved and placed where ever you want it to go. REMBEMBER TO MOVE the MAP THAT HAS THE YELLOW SYMBOL! IF THATS NOT MOVED YOUR NOT MOVING THE CORRECT ONE AND THE ANIMATION WILL BE PLAYED WHERE YOU PUT IT DOWN.