Attemoted_some_feeld_animations
There is small_animations_on feelds, & carrots. don't know what scrips are supossed to be on them. So they don't work, besides playing the idle. They don't have the right matterial either, beside the carrots. So, when things like scrips and coliders and things have been placed you should drag that into a folder of your choice and delete the one I have, so that we have one complete ready to drop into the game version.